+++ demo/meeting/src/containers/MeetingDetails/index.tsx
@@ -6,14 +6,16 @@ import React from 'react';
 import {
   Flex,
   Heading,
-  PrimaryButton
+  PrimaryButton,
+  useMeetingManager
 } from 'amazon-chime-sdk-component-library-react';
 
 import { useAppState } from '../../providers/AppStateProvider';
 import { StyledList } from './Styled';
 
 const MeetingDetails = () => {
-  const { meetingId, region, toggleTheme, theme } = useAppState();
+  const { meetingId, toggleTheme, theme } = useAppState();
+  const manager = useMeetingManager();
 
   return (
     <Flex container layout="fill-space-centered">
@@ -28,7 +30,7 @@ const MeetingDetails = () => {
           </div>
           <div>
             <dt>Hosted region</dt>
-            <dd>{region}</dd>
+            <dd>{manager.meetingRegion}</dd>
           </div>
         </StyledList>
         <PrimaryButton