+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+- Fixed reversed labels for mute/unmute control in `AudioInputControl`
 
 ## [1.0.3] - 2020-08-04