+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Storybook to 5.3
 - Improve docs for SDK components
 - Portrait video support, minor style fixes
+- [Demo] Update demo dependencies
 
 ### Removed