+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2020-08-14
 
 ### Added
 
 - Add Github documentation link to README
 - Add SDK log level selection ability to demo and library
+- Add PostLogger support to MeetingProvider
 
 ### Changed
 
@@ -20,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Storybook to 5.3
 - Improve docs for SDK components
 - Portrait video support, minor style fixes
+- [Demo] Update demo dependencies
+- [Demo] Default to nearest available region
 
 ### Removed